refactor(server): relocate background-agent logic out of server.go
Move the background-agent execution + lifecycle functions
(executeBackgroundDelegatedTask, the completion-polling loop, batched
start/completion notification handling, and auto-notification message
builders — ~1,550 lines) from the 11.7k-line server.go into the
existing background_agents.go, where the BackgroundAgent types and
registry already live. Pure relocation: same package, no behavior
change, imports resolved by goimports.

server.go 11,738 → 10,189 lines. fmt/vet clean, cmd/server tests pass.
